pyproject.toml
[tool.poetry]
name = "mysite"
version = "0.1.0"
description = ""
authors = ["Ankit Kumar Singh <ankit.singh@grofers.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.5"
dj-database-url = "^2.1.0"
psycopg2-binary = "^2.9.7"
gunicorn = "^21.2.0"
whitenoise = {extras = ["brotli"], version = "^6.5.0"}
black = "^23.9.1"
django-rest-framework = "^0.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
